+++ fenestra-kit/src/lib.rs
@@ -6,6 +6,7 @@
 #![warn(missing_docs)]
 
 mod demo;
+mod font_features;
 mod gallery;
 pub mod icons;
 mod specimen;
@@ -13,6 +14,7 @@ mod type_specimen;
 mod widgets;
 
 pub use demo::{ai_chat, editor_panel, holy_grail, poster, scroll_demo};
+pub use font_features::font_feature_specimen;
 pub use gallery::{gallery_controls, gallery_display};
 pub use specimen::specimen;
 pub use type_specimen::type_specimen;